requirements.txt
flask
pdfplumber
sentence-transformers
faiss-cpu
numpy
requests